+++ config/initializers/container.rb
@@ -82,6 +82,9 @@ module Charkeeper
     register('commands.characters_context.daggerheart.homebrew.add_item') {
       CharactersContext::Daggerheart::Homebrew::AddItemCommand.new
     }
+    register('commands.characters_context.dnd2024.homebrew.add_item') {
+      CharactersContext::Dnd2024::Homebrew::AddItemCommand.new
+    }
 
     register('commands.image_processing.attach_avatar_by_file') { ImageProcessingContext::AttachAvatarByFileCommand.new }
     register('commands.image_processing.attach_avatar_by_url') { ImageProcessingContext::AttachAvatarByUrlCommand.new }