+++ config/routes.rb
@@ -106,6 +106,7 @@ Rails.application.routes.draw do
         resources :rest, only: %i[create], module: 'characters'
         resources :craft, only: %i[index create], module: 'characters'
         resources :talents, only: %i[index create], module: 'characters'
+        resources :homebrew_items, only: %i[create], module: 'characters'
       end
       resources :spells, only: %i[index]
     end